Update to definitions v5.0.0
The v5.0.0 definitions include a breaking format change related to
the 'year_ranges' feature. This PR includes all required code changes
to consume this updated format.

While holiday behavior has not been changed this commit introduces
a restriction to only allow one 'year_range' operator at a time.
This was the defacto way that this feature was used in the past but
this change makes it explicit.

If definitions somehow violate this rule (the defs have their own
similar validation to prevent this) then we will now throw an explicit
error instead of silently taking the first operator to apply.