Proper cache validation
Currently if you call `on` on a date that doesn't havea a holiday, cache does properly return an empty array. Because of the simplistic check it's misinterpreted as no cache data. Which renders cache useless if you check a lot of empty days. Instead it should be checked for `nil` to distinguish empty cache and missing holidays.